kernel,dev: share gateway transform http handler with dev server
